[Bronze V] Title: 이상한 기호, Time: 68 ms, Memory: 11496 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -11,7 +11,7 @@ public class Main {
 		System.out.println(op(A, B));
 	}
 	
-	static int op(int A, int B) {
-		return (A + B)*(A - B);
+	static long op(int A, int B) {
+		return (long)(A + B)*(A - B);
 	}
 }
